+++ core/src/main/kotlin/net/corda/core/node/NetworkParameters.kt
@@ -2,8 +2,10 @@ package net.corda.core.node
 
 import net.corda.core.CordaRuntimeException
 import net.corda.core.KeepForDJVM
+import net.corda.core.contracts.ContractClassName
 import net.corda.core.crypto.toStringShort
 import net.corda.core.identity.Party
+import net.corda.core.internal.VisibleForTesting
 import net.corda.core.internal.requirePackageValid
 import net.corda.core.node.services.AttachmentId
 import net.corda.core.serialization.CordaSerializable
@@ -185,7 +187,17 @@ data class NetworkParameters(
     /**
      * Returns the public key of the package owner of the [contractClassName], or null if not owned.
      */
-    fun getOwnerOf(contractClassName: String): PublicKey? = this.packageOwnership.filterKeys { packageName -> owns(packageName, contractClassName) }.values.singleOrNull()
+    @VisibleForTesting
+    internal fun getPackageOwnerOf(contractClassName: ContractClassName): PublicKey? = this.packageOwnership.filterKeys { packageName -> owns(packageName, contractClassName) }.values.singleOrNull()
+
+    /**
+     * Returns the public key of the package owner if any of [contractClassName] match, or null if not owned.
+     */
+    @VisibleForTesting
+    internal fun getPackageOwnerOf(contractClassNames: Set<ContractClassName>): PublicKey? {
+        val ownerKeys = contractClassNames.map { getPackageOwnerOf(it) }
+        return ownerKeys.find { it != null }
+    }
 
     /**
      * Returns true if the only properties changed in [newNetworkParameters] are [AutoAcceptable] and not